Added documentation and extra operators
