--- Assets/LeapMotion/Core/Scripts/Query/DirectOperators.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: a1b69291f1f25af46b877fa26e25c273
-timeCreated: 1521663092
-licenseType: Pro
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
