+++ Assets/LeapMotion/Core/Scripts/Query/QueryValueGenerators.cs
